Refactor local_sgd integration tests
